float to double

+++ TP1/include/Cena.hpp
@@ -14,8 +14,8 @@ public:
 
     //kinda getters and setters
     vector<objeto>& getCena();
-    void addObject(const float &x, const float &y, const float &largura);
-    void addMovement(const int &object, const int &tempo, const float &x, const float &y);
+    void addObject(const double &x, const double &y, const double &largura);
+    void addMovement(const int &object, const int &tempo, const double &x, const double &y);
 
     // Sorting methods
     void cenaSortTime(const int &time);
